Implement warm-start optimization for Taichi gem solver
- Added genome_hint_allocation field to store best gem allocations
- Implemented local_search_from_hint kernel for warm-start refinement
- Added kernels for storing and inheriting hints during GA
- Updated ga_evaluate_population to support warm-start via use_hints flag
- Integrated warm-start logic into genetic.py GA loop
- Exported new kernels and API functions

+++ gear_optimizer/solver/taichi_gem/api/initialization.py
@@ -168,7 +168,7 @@ def _ensure_batch_staging():
         "work_items": np.zeros((MAX_WORK_ITEMS, 8), dtype=np.int32),
         "fever_masks": np.zeros((MAX_WORK_ITEMS, MAX_HEAD_NOTES), dtype=np.int8),
         # [pp, cm, fm, p_val, s_val, ft, ff]
-        "genome_base_stats": np.zeros((MAX_GENOMES, 7), dtype=np.int32),
+        "genome_base_stats": np.zeros((MAX_GENOMES, 7), dtype=np.int16),
     }
     return _BATCH_STAGING
 
@@ -182,7 +182,7 @@ def _ensure_mega_staging():
         "work_items": np.zeros((MAX_WORK_ITEMS, 8), dtype=np.int32),
         "fever_masks": np.zeros((MAX_WORK_ITEMS, MAX_HEAD_NOTES), dtype=np.int8),
         # [pp, cm, fm, p_val, s_val, ft, ff]
-        "genome_base_stats": np.zeros((MAX_GENOMES, 7), dtype=np.int32),
+        "genome_base_stats": np.zeros((MAX_GENOMES, 7), dtype=np.int16),
     }
     return _MEGA_STAGING
 
@@ -200,7 +200,7 @@ def _ensure_parallel_staging():
     _PARALLEL_STAGING = {
         # Per-genome stats (uploaded once per call)
         # [pp, cm, fm, p_val, s_val, ft, ff]
-        "genome_base_stats": np.zeros((MAX_GENOMES, 7), dtype=np.int32),
+        "genome_base_stats": np.zeros((MAX_GENOMES, 7), dtype=np.int16),
 
         # Per-work-item buffers (reused per chunk)
         # [budget, count_fever, count_normal, ft_gems, ff_gems, head_len, genome_id, song_slot]
